fixed error in alembic migration

+++ migrations/versions/d975ea83b712_add_seft_ci_file_location.py
@@ -18,7 +18,7 @@ depends_on = None
 
 def upgrade():
     op.add_column("instrument", sa.Column("file_location", String(255)), schema="ras_ci")
-    op.add_column("instrument", sa.Column("file_length", Integer(10)), schema="ras_ci")
+    op.add_column("instrument", sa.Column("file_length", Integer), schema="ras_ci")
 
 
 def downgrade():
